sua lai phan add bill

+++ src/main/java/com/example/tmdt/model/Product.java
@@ -30,7 +30,7 @@ public class Product {
     @Min(value = 1)
     private Double promotion;
     @Column(columnDefinition = "integer default 0")
-    private Integer count;
+    private Double count;
     @Column(nullable = true)
     private Integer status;
     @OneToMany(cascade = CascadeType.ALL)
